+++ example/data/lib/features/catalog/local/models/catalog_box.dart
@@ -1,0 +1,9 @@
+class CatalogBox {
+  const CatalogBox({
+    this.id = 0,
+    this.remoteId = '',
+  });
+
+  final int id;
+  final String remoteId;
+}