+++ example/data/lib/features/catalog/remote/catalog_remote_data_source.dart
@@ -1,0 +1,5 @@
+import 'models/catalog_dto.dart';
+
+abstract interface class CatalogRemoteDataSource {
+  Future<List<CatalogDto>> getItems();
+}