+++ example/domain/lib/features/catalog/entities/catalog_entity.dart
@@ -1,0 +1,7 @@
+class CatalogEntity {
+  const CatalogEntity({
+    required this.remoteId,
+  });
+
+  final String remoteId;
+}