+++ example/domain/lib/features/catalog/repositories/catalog_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/catalog_entity.dart';
+
+abstract interface class CatalogRepository {
+  Future<List<CatalogEntity>> getCatalogList();
+}