+++ lib/src/operation_kind.dart
@@ -1,0 +1,8 @@
+enum OperationKind {
+  remote,
+  local,
+  cached;
+
+  bool get includesRemote => this == remote || this == cached;
+  bool get includesLocal => this == local || this == cached;
+}